rename: 대소문자가 바뀐 파일명, 폴더명이 인식되지 않아 임의 수정 (#101)
